Refactor status manager into modular components
