+++ V2_COMPLIANCE_PROGRESS_TRACKER.md
@@ -169,6 +169,13 @@
 - **Completion Date**: 2025-08-24
 - **Summary**: Refactored from 722 to 55 lines by extracting manager, AI processor, coordinator, and config modules. New orchestrator coordinates components and maintains functionality.
 
+### MODERATE-017: Status Manager Modularization ✅
+- **File**: `src/core/status_manager.py`
+- **Status**: Completed
+- **Assigned To**: Agent-1
+- **Completion Date**: 2025-08-24
+- **Summary**: Extracted tracking, reporting, and configuration into dedicated modules and replaced original file with a lightweight orchestrator.
+
 ## 📋 AVAILABLE CONTRACTS FOR CLAIMING
 
 > **Note:** Remaining contract descriptions include current line counts for context only. There is no strict LOC target—deliver clean, production-ready, tested modules that honor SRP and SOLID principles.
